Make use of the easyfxml boilerplate classes just added
