+++ src/main/java/moe/lyrebird/Lyrebird.java
@@ -1,43 +1,19 @@
 package moe.lyrebird;
 
-import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.context.annotation.Import;
-import moe.tristan.easyfxml.spring.SpringContext;
-import lombok.extern.slf4j.Slf4j;
-import moe.lyrebird.view.GuiBootstraper;
-
-import javafx.application.Application;
-import javafx.stage.Stage;
+import moe.tristan.easyfxml.spring.application.FxSpringApplication;
+import moe.tristan.easyfxml.spring.application.FxSpringContext;
 
 /**
  * Main application entry point.
  */
-@Slf4j
 @SpringBootApplication
-@Import(SpringContext.class)
-public class Lyrebird extends Application {
-    private ConfigurableApplicationContext context;
+@Import(FxSpringContext.class)
+public class Lyrebird extends FxSpringApplication {
 
-    @Override
-    public void init() {
-        this.context = SpringApplication.run(Lyrebird.class);
-    }
-
-    @Override
-    public void start(final Stage primaryStage) {
-        this.context.getBean(GuiBootstraper.class).startGui(primaryStage);
-    }
-
-    @Override
-    public void stop() {
-        this.context.stop();
-    }
-
-    public static void main(final String... args) {
-        Application.launch(args);
+    public static void main(final String[] args) {
+        launch(args);
     }
 
 }
-
