+++ src/main/resources/style/palette.scss
@@ -1,0 +1,15 @@
+/* Coolors Exported Palette - coolors.co/e03616-33658a-2f4858-fdfffc-235789 */
+
+/* HSL */
+$color1: hsla(10%, 82%, 48%, 1);
+$color2: hsla(206%, 46%, 37%, 1);
+$color3: hsla(203%, 30%, 26%, 1);
+$color4: hsla(100%, 100%, 99%, 1);
+$color5: hsla(209%, 59%, 34%, 1);
+
+/* RGB */
+$color1: rgba(224, 54, 22, 1);
+$color2: rgba(51, 101, 138, 1);
+$color3: rgba(47, 72, 88, 1);
+$color4: rgba(253, 255, 252, 1);
+$color5: rgba(35, 87, 137, 1);
