+++ src/main/resources/vendors/liberation_sans/AUTHORS
@@ -1,0 +1,10 @@
+Font Designer:
+  Ascender Corporation.
+  http://www.ascendercorp.com/
+
+Original Designer:
+  Steve Matteson
+  http://www.ascendercorp.com/typedesigners.html
+
+Current Maintainer:
+  Caius Chance <cchance@redhat.com>
