+++ src/main/resources/vendors/liberation_sans/README
@@ -1,0 +1,68 @@
+1. What's this?
+===============
+
+The Liberation Fonts are intended to be open font replacements for the three
+most commonly used fonts on Microsoft systems: Times New Roman, Arial, and
+Courier New.
+
+
+2. Requirements
+===============
+
+  * fontforge.
+    - http://fontforge.sourceforge.net
+
+
+3. Install
+==========
+
+3.1 Decompress tarball
+
+  You can extract the files by following command:
+
+    $ tar zxvf liberation-fonts-[VERSION].tar.gz
+
+3.2 Build from the source
+
+  Change into directory liberation-fonts-[VERSION]/ and build from sources by 
+  following commands:
+
+    $ cd liberation-fonts-[VERSION]
+    $ make
+
+  The built font files will be available in 'build' directory.
+
+3.3 Install to system
+
+  You can manually install the fonts by copying them to ~/.fonts for a
+  particular user, or to /usr/share/fonts/truetype/liberation for system-wide
+  availability
+
+
+4. Usage
+========
+
+The fonts should be installed and detected by the system after installation. 
+
+
+5. License
+==========
+
+Read the GPL v2 file 'COPYING' and the Liberation Fonts EULA file 'License.txt'
+for more information.
+
+6. Credits
+==========
+
+Special thanks to all liberation-fonts developers and users!
+
+ * Caius Chance <cchance at redhat.com>
+   - Current project maintainer.
+
+ * Mark Webbink <mwebbink at redhat.com>
+   - Coordinator.
+
+ * Steve Matteson
+   - Ascender Corporation <http://www.ascendercorp.com>
+
+ * And, all other contributors.
